refactor: 레이아웃 관련 래퍼 컴포넌트 common으로 이동
